V1.8 - Refactoring image loading manually/automatically and moved manual dialog to seperate class

+++ droparea.py
@@ -2,6 +2,7 @@ import os
 from PyQt5.QtWidgets import *
 from PyQt5.QtGui import *
 from PyQt5.QtCore import *
+from addtabdialog import AddTabDialog
 import constants
 
 MIN_FOR_CLOSE = 100
@@ -38,11 +39,6 @@ class DropArea(QLabel):
         self.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Maximum)
         self.setCursor(Qt.PointingHandCursor)
         self.setMouseTracking(True)
-
-        button = QPushButton("x", self)
-        button.move(0, 0)
-        button.show()
-        button.clicked.connect(lambda: self.tabDeleted.emit(self))
 
         self.tabNumber = tabNumber
         self.clear()
@@ -92,44 +88,8 @@ class DropArea(QLabel):
         # OPEN TAB IN BROWSER OR IF EMPTY ENTER URL MANUALLY
         else: 
             if self.taken: QDesktopServices.openUrl(QUrl(self.url))
-            else:
-                dialog = QDialog()
-                layout = QVBoxLayout()
-                label = QLabel("Enter URL: ")
-                label.setAlignment(Qt.AlignCenter)
-                textEdit = QLineEdit("https://www.google.com")
-                textEdit.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
-                textEdit.setFrame(QFrame.NoFrame)
-                button = QPushButton("Add URL")
+            else: dialog = AddTabDialog(self)
 
-                def dialogButton():
-                    self.loadFromUrl(textEdit.text())
-                    dialog.deleteLater()
-                button.clicked.connect(dialogButton)
-
-                urlList = QListWidget()
-                path = os.path.join(os.getcwd(), '.tabs.txt')
-                if os.path.isfile(path):
-                    if os.path.getsize(path):
-                        with open('.tabs.txt', 'r') as f:
-                            lines = f.readlines()
-                            for line in lines:
-                                cur = line.split('\n')[0]
-                                # GET URL FROM LINE (URL IMAGEPATH)
-                                current = QListWidgetItem(cur.split(' ')[0])
-                                urlList.addItem(current)
-
-                urlList.currentItemChanged.connect(lambda current, previous: textEdit.setText(current.text()))
-
-                layout.addWidget(label)
-                layout.addWidget(textEdit)
-                layout.addWidget(urlList)
-                layout.addWidget(button)
-                dialog.setWindowModality(Qt.ApplicationModal)
-                dialog.setWindowTitle("Add Manually")
-                dialog.setLayout(layout)
-                dialog.exec_()
-        
         if self.taken:
             self.setFrameStyle(QFrame.NoFrame)
             self.setLineWidth(2)
@@ -158,7 +118,8 @@ class DropArea(QLabel):
 
             ### HANDLE EXCEPTIONS WERE PAGES ARE BLOCKED OR NETWORK DOESNT WORK ETC
             ### WHEN GOING BACK ONLINE LOAD IMAGE??
-            self.downloadImage()
+            self.downloadImage(self.url)
+            self.setLocalPixmap()
             self.saveTab()
             event.acceptProposedAction()
             # CONNECTED TO ADD TAB TO LIST
@@ -182,23 +143,18 @@ class DropArea(QLabel):
                     return QSize(width, width*self.aspectRatio)
         return QSize(self.minWidth, self.minHeight)
 
+    def addCloseButton(self):
+        button = QPushButton("x", self)
+        button.move(0, 0)
+        button.show()
+        button.clicked.connect(lambda: self.tabDeleted.emit(self))
+
     # LOAD TAB PIXMAP AND SET CLASS VARIABLES
     def load(self, url, imagePath, addTabAfter):
         self.taken = True
         self.url = url
-        self.tabAdded.emit(self, self.tabNumber)
-        self.imagePath = imagePath
-        self._pixmap = QPixmap(os.path.join(self.imageFolder, imagePath)).scaled(self.sizeHint(), Qt.KeepAspectRatio)
-        self.setPixmap(self._pixmap)
-        # CONNECTED TO NEWTAB
-        if addTabAfter: self.imageLoaded.emit(False)
 
-    # LOAD FUNCTION OVERLOAD WITH ONLY URL FOR MANUALLY ADDNG TAB
-    def loadFromUrl(self, url):
-        self.taken = True
-        self.url = url
-        self.imagePath = str(url.replace('/','') + '.png')
-
+        loaded = False
         if os.path.isfile('.tabs.txt') and os.path.getsize('.tabs.txt') > 0:
             with open('.tabs.txt', 'r') as f:
                 lines = f.readlines()
@@ -207,39 +163,68 @@ class DropArea(QLabel):
                     curUrl = cur.split(' ')[0]
                     imagePath = cur.split(' ')[1]
                     if curUrl == url:
-                        if not os.path.isfile(os.path.join(self.imageFolder, imagePath[1:])):
-                            os.rename(os.path.join(self.imageFolder, imagePath), os.path.join(self.imageFolder, imagePath[1:]))
-                        self._pixmap = QPixmap(os.path.join(self.imageFolder, imagePath[1:]), '1')
-                        self.imagePath = imagePath[1:]
-                        self.url = url
-                        self.setPixmap(self._pixmap.scaled(self.sizeHint(), Qt.KeepAspectRatio))
-                        self.imageLoaded.emit(False)
-        else:
-            if not self.checkDuplicate(url):
-                self.driver.get(url)
-                self.driver.save_screenshot(os.path.join(self.imageFolder, self.imagePath))
-                self.setLocalPixmap()
+                        if not os.path.isfile(os.path.join(self.imageFolder, imagePath[1:])): os.rename(os.path.join(self.imageFolder, imagePath), os.path.join(self.imageFolder, imagePath[1:]))
+                        self.imagePath = imagePath[1:]                        
+                        loaded = True
+        if not loaded:
+            self.imagePath = imagePath
+        
+        # CONNECTED TO NEWTAB
+        if addTabAfter: self.imageLoaded.emit(False)
+        self._pixmap = QPixmap(os.path.join(self.imageFolder, imagePath))
+        self.setPixmap(self._pixmap.scaled(self.sizeHint(), Qt.KeepAspectRatio))
         self.tabAdded.emit(self, self.tabNumber)
-        self.saveTab()
+        self.addCloseButton()
+
+    # LOAD FUNCTION OVERLOAD WITH ONLY URL FOR MANUALLY ADDNG TAB
+    #def loadFromUrl(self, url):
+    #    self.taken = True
+    #    self.url = url
+    #    self.imagePath = str(url.replace('/','') + '.png')
+
+    #    loaded = False
+    #    if os.path.isfile('.tabs.txt') and os.path.getsize('.tabs.txt') > 0:
+    #        with open('.tabs.txt', 'r') as f:
+    #            lines = f.readlines()
+    #            for line in lines:
+    #                cur = line.split('\n')[0]
+    #                curUrl = cur.split(' ')[0]
+    #                imagePath = cur.split(' ')[1]
+    #                if curUrl == url:
+    #                    if not os.path.isfile(os.path.join(self.imageFolder, imagePath[1:])):
+    #                        os.rename(os.path.join(self.imageFolder, imagePath), os.path.join(self.imageFolder, imagePath[1:]))
+    #                    self._pixmap = QPixmap(os.path.join(self.imageFolder, imagePath[1:]), '1')
+    #                    self.imagePath = imagePath[1:]
+    #                    self.url = url
+    #                    self.setPixmap(self._pixmap.scaled(self.sizeHint(), Qt.KeepAspectRatio))
+    #                    self.imageLoaded.emit(False)
+    #                    loadded = True
+    #                    print('Loaded URL from file:::')
+    #    if not loaded:
+    #        if not self.checkDuplicate(url):
+    #            self.driver.get(url)
+    #            self.driver.save_screenshot(os.path.join(self.imageFolder, self.imagePath))
+    #            self.setLocalPixmap()
+    #            print('Loaded URL from driver:::')
+    #    self.tabAdded.emit(self, self.tabNumber)
+    #    self.saveTab()
+    #    self.addCloseButton()
 
     # DOWNLOAD IMAGE FROM DRIVER
-    def downloadImage(self):
-        self.imagePath = str(self.url).replace('/', '') + '.png'
-        if not self.checkDuplicate(self.url):
-            self.driver.get(self.url)
+    def downloadImage(self, url):
+        self.imagePath = str(url).replace('/', '') + '.png'
+        if not self.checkDuplicate(url):
+            self.driver.get(url)
             self.driver.save_screenshot(os.path.join(self.imageFolder, self.imagePath))
         
         if self.checkArchivedTabsExists():
             with open('.tabs.txt', 'r') as f:
                 lines = f.readlines()
                 for line in lines:
-                    if line.split(' ')[0] == self.url:
+                    if line.split(' ')[0] == url:
                         imagePath = line.split('\n')[0].split(' ')[1]
                         os.rename(os.path.join(os.imageFolder, imagePath), os.path.join(os.imageFolder, imagePath[1:]))
                         self.imagePath = imagePath[1:]
-                        #self._pixmap(os.path.join(self.imageFolder, imagePath[1:]))
-                        #self.setPixmap(self._pixmap.scaled(self.sizeHint(), Qt.KeepAspectRatio, Qt.SmoothTransformation))
-        self.setLocalPixmap()
     
     # SET PIXMAP FROM IMAGEPATH MEMBER VARIABLE AND ADD NEW TAB AFTER
     def setLocalPixmap(self):
@@ -250,7 +235,9 @@ class DropArea(QLabel):
     # CHECK IF FILE EXISTS AND IF ITS NOT EMPTY
     def checkArchivedTabsExists(self):
         exists = os.path.isfile(os.path.join(os.getcwd(), '.tabs.txt'))
-        notEmpty = True if (os.path.getsize(os.path.join(os.getcwd(), '.tabs.txt')) > 0) else False
+        notEmpty = False
+        if exists:
+            notEmpty = True if (os.path.getsize(os.path.join(os.getcwd(), '.tabs.txt')) > 0) else False
         return (exists and notEmpty)
 
     def checkDuplicate(self, url):
